Move Ensembl ID to gene name function into utils
